refactor(BA-6244): consolidate keypair-by-resource-policy action and use proper types
- Move SearchKeypairsByResourcePolicyAction/Result into keypair_ops.py instead of
  a standalone module, alongside the other keypair search actions.
- Use the existing KeypairResourcePolicyNotFound error for the scope existence
  check instead of the generic ObjectNotFound.
- Add ScopeType.KEYPAIR_RESOURCE_POLICY and use it for the action's scope_type()
  (value matches RBACElementType so scope_type().to_element() resolves correctly).
